@@ -5,49 +5,50 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.11",
-    "@angular/common": "~10.0.11",
-    "@angular/compiler": "~10.0.11",
-    "@angular/core": "~10.0.11",
-    "@angular/forms": "~10.0.11",
-    "@angular/localize": "^10.0.11",
-    "@angular/platform-browser": "~10.0.11",
-    "@angular/platform-browser-dynamic": "~10.0.11",
-    "@angular/router": "~10.0.11",
-    "@fortawesome/fontawesome-free": "^5.12.0",
-    "@ng-bootstrap/ng-bootstrap": "^6.0.0",
-    "@okta/okta-angular": "^3.0.1",
-    "@okta/okta-signin-widget": "^5.2.0",
-    "bootstrap": "^4.4.1",
-    "rxjs": "~6.5.4",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "@angular/animations": "^14.0.0",
+    "@angular/common": "^14.0.0",
+    "@angular/compiler": "^14.0.0",
+    "@angular/core": "^14.0.0",
+    "@angular/forms": "^14.0.0",
+    "@angular/platform-browser": "^14.0.0",
+    "@angular/platform-browser-dynamic": "^14.0.0",
+    "@angular/router": "^14.0.0",
+    "@fortawesome/fontawesome-free": "^6.1.2",
+    "@ng-bootstrap/ng-bootstrap": "^13.0.0",
+    "@okta/okta-angular": "^5.2.0",
+    "@okta/okta-auth-js": "^6.4.0",
+    "@okta/okta-signin-widget": "^6.2.0",
+    "bootstrap": "^5.2.0",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.2",
-    "@angular/cli": "^10.0.7",
-    "@angular/compiler-cli": "~10.0.11",
-    "@angular/language-service": "~10.0.11",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "@angular-devkit/build-angular": "^14.0.7",
+    "@angular-eslint/builder": "14.0.2",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/schematics": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
+    "@angular/cli": "~14.0.7",
+    "@angular/compiler-cli": "^14.0.0",
+    "@angular/localize": "^14.1.1",
+    "@types/jasmine": "~4.0.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
+    "eslint": "^8.18.0",
+    "jasmine-core": "~4.1.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.9.7"
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.7.2"
   }
 }
