@@ -1,7 +1,9 @@
 export class Address {
-    street: string;
-    city: string;
-    state: string;
-    country: string;
-    zipCode: string;
+    
+    constructor(public street: string,
+                public city: string,
+                public state: string,
+                public country: string,
+                public zipCode: string) { }
+
 }
