@@ -1,7 +1,10 @@
 export class OrderHistory {
-    id: string;
-    orderTrackingNumber: string;
-    totalPrice: number;
-    totalQuantity: number;
-    dateCreated: Date;
+
+    constructor(public id: string,
+                public orderTrackingNumber: string,
+                public totalPrice: number,
+                public totalQuantity: number,
+                public dateCreated: Date) {
+
+                }
 }
