@@ -1,15 +1,8 @@
-import { CartItem } from './cart-item';
-
 export class OrderItem {
-    imageUrl: string;
-    unitPrice: number;
-    quantity: number;
-    productId: string;
 
-    constructor(cartItem: CartItem) {
-        this.imageUrl = cartItem.imageUrl;
-        this.quantity = cartItem.quantity;
-        this.unitPrice = cartItem.unitPrice;
-        this.productId = cartItem.id;
-    }
+    constructor(public imageUrl: string,
+                public unitPrice: number,
+                public quantity: number,
+                public productId: string) { }
+
 }
