@@ -1,4 +1,6 @@
 export class Order {
-    totalQuantity: number;
-    totalPrice: number;
+    
+    constructor (public totalQuantity: number,
+                 public totalPrice: number) { }
+                 
 }
