@@ -1,4 +1,7 @@
 export class ProductCategory {
-    id: number;
-    categoryName: string;
+    
+    constructor(public id: number,
+                public categoryName: string) {
+
+    }
 }
