@@ -4,9 +4,9 @@ import { Order } from './order';
 import { OrderItem } from './order-item';
 
 export class Purchase {
-    customer: Customer;
-    shippingAddress: Address;
-    billingAddress: Address;
-    order: Order;
-    orderItems: OrderItem[]; 
+    customer!: Customer;
+    shippingAddress!: Address;
+    billingAddress!: Address;
+    order!: Order;
+    orderItems!: OrderItem[]; 
 }
