@@ -1,4 +1,6 @@
 export class State {
-    id: number;
-    name: string;
+    
+    constructor(public id: number,
+                public name: string) { }
+
 }
