@@ -1,5 +1,6 @@
-import { Component, OnInit } from '@angular/core';
-import { OktaAuthService } from '@okta/okta-angular';
+import { Component, Inject, OnInit } from '@angular/core';
+import { OktaAuthStateService, OKTA_AUTH } from '@okta/okta-angular';
+import { OktaAuth } from '@okta/okta-auth-js';
 
 @Component({
   selector: 'app-login-status',
@@ -9,33 +10,33 @@ import { OktaAuthService } from '@okta/okta-angular';
 export class LoginStatusComponent implements OnInit {
 
   isAuthenticated: boolean = false;
-  userFullName: string;
+  userFullName: string = '';
 
   storage: Storage = sessionStorage;
 
-  constructor(private oktaAuthService: OktaAuthService) { }
+  constructor(private oktaAuthService: OktaAuthStateService,
+    @Inject(OKTA_AUTH) private oktaAuth: OktaAuth) { }
 
   ngOnInit(): void {
 
     // Subscribe to authentication state changes
-    this.oktaAuthService.$authenticationState.subscribe(
+    this.oktaAuthService.authState$.subscribe(
       (result) => {
-        this.isAuthenticated = result;
+        this.isAuthenticated = result.isAuthenticated!;
         this.getUserDetails();
       }
     );
-    
   }
-
+  
   getUserDetails() {
     if (this.isAuthenticated) {
 
       // Fetch the logged in user details (user's claims)
       //
       // user full name is exposed as a property name
-      this.oktaAuthService.getUser().then(
+      this.oktaAuth.getUser().then(
         (res) => {
-          this.userFullName = res.name;
+          this.userFullName = res.name as string;
 
           // retrieve the user's email from authentication response
           const theEmail = res.email;
@@ -49,6 +50,7 @@ export class LoginStatusComponent implements OnInit {
 
   logout() {
     // Terminates the session with Okta and removes current tokens.
-    this.oktaAuthService.signOut();
+    this.oktaAuth.signOut();
   }
+
 }
