@@ -21,7 +21,7 @@ export class OrderHistoryComponent implements OnInit {
   handleOrderHistory() {
 
     // read the user's email address from browser storage
-    const theEmail = JSON.parse(this.storage.getItem('userEmail'));
+    const theEmail = JSON.parse(this.storage.getItem('userEmail')!);
 
     // retrieve data from the service
     this.orderHistoryService.getOrderHistory(theEmail).subscribe(
