@@ -1,8 +1,8 @@
 export default {
 
     oidc: {
-        clientId: '<<ADD-YOUR-CLIENT-ID>>>',
-        issuer: 'https://<<ADD-YOUR-DEV-DOMAIN>>/oauth2/default',
+        clientId: '<<UPDATE-WITH-YOUR-APP-CLIENT-ID>>',
+        issuer: 'https://<<UPDATE-WITH-YOUR-DEV-DOMAIN>>/oauth2/default',
         redirectUri: 'http://localhost:4200/login/callback',
         scopes: ['openid', 'profile', 'email']
     }
